FIX: ALSC: fix wrong number for free&malloc when update iq;
When update calib, should free original data,
but the number of ptr to free has been update with new iq,
the actual number of ptr to free is from old iq,
then it will cause out of range.

This patch save the old number to be used when free data.

Change-Id: If6291d191aec5dade94876dfeced84cfd677e28b
Signed-off-by: archer.wu <archer.wu@rock-chips.com>

Signed-off-by: dalon.zhang <dalon.zhang@rock-chips.com>
Change-Id: Id544d599c9c70b83fe141d1d238a787a9cb56af8